Rename loginpanel to loginform
Run prettier and move constants to `constants.js`
